--- src/pages/HomePage.tsx
@@ -1,5 +1,0 @@
-function HomePage() {
-  return <div>HomePage</div>;
-}
-
-export default HomePage;
